LPD-49114 Move util from structure_build scope to a more generic one
