+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/util/getImage.ts
